move renderVideoOrImage to DetailsPageUtils
